Alters setup so that a Next.js server is deployed as a Cloud Function
Before serverless Next.js pages were deployed as Cloud Functions and the appropriate URLs were rewritten to them in an effort to integrate the serverless Next.js build with Firebase. Now a single Cloud Function is deployed and all URLs are rewritten to it and the Cloud Function handle requests using a Next.js server.